Include specific line and direction
