+++ packages/app/src/api/client.ts
@@ -1,27 +1,28 @@
 import { z } from "zod";
 
 import { config } from "../config";
-import { recentResponse } from "./mockresponse";
 
 export const reportSchema = z
   .object({
     timestamp: z.string().transform((value) => new Date(value)),
+    line: z.string(),
+    direction: z.object({
+      id: z.string(),
+      name: z.string(),
+    }),
     station: z.object({
       id: z.string(),
     }),
   })
-  .transform(({ timestamp, station }) => ({
-    timestamp,
+  .transform(({ station, ...rest }) => ({
+    ...rest,
     stationId: station.id,
   }));
 
 export type Report = z.infer<typeof reportSchema>;
 
 const getReports = async (): Promise<Report[]> => {
-  if (__DEV__) {
-    return reportSchema.array().parse(recentResponse);
-  }
-  const response = await fetch(`${config.FF_API_BASE_URL}/recent`);
+  const response = await fetch(`${config.FF_API_BASE_URL}/basics/recent`);
   const data = await response.json();
 
   return reportSchema.array().parse(data);
@@ -45,13 +46,16 @@ type PostReport = {
 };
 
 const postReport = async (report: PostReport) => {
-  const response = await fetch(`${config.FF_API_BASE_URL}/newInspector`, {
-    method: "POST",
-    body: JSON.stringify(report),
-    headers: {
-      "Content-Type": "application/json",
-    },
-  });
+  const response = await fetch(
+    `${config.FF_API_BASE_URL}/basics/newInspector`,
+    {
+      method: "POST",
+      body: JSON.stringify(report),
+      headers: {
+        "Content-Type": "application/json",
+      },
+    }
+  );
 
   return response.json();
 };
