--- packages/app/src/api/mockresponse.ts
@@ -1,86 +1,0 @@
-export const recentResponse = [
-  {
-    timestamp: "2024-03-25T19:40:22.786354Z",
-    station: {
-      id: "U-Kbo",
-      name: "Kottbusser Tor",
-      coordinates: {
-        latitude: 52.4990174,
-        longitude: 13.4184555,
-      },
-    },
-    direction: {
-      id: "",
-      name: "",
-      coordinates: {
-        latitude: 0,
-        longitude: 0,
-      },
-    },
-    line: "",
-    isHistoric: true,
-  },
-  {
-    timestamp: "2024-03-25T19:27:27.234033Z",
-    station: {
-      id: "SU-Ts",
-      name: "Tempelhof",
-      coordinates: {
-        latitude: 52.4709155,
-        longitude: 13.3847575,
-      },
-    },
-    direction: {
-      id: "U-ATg",
-      name: "Alt-Tegel",
-      coordinates: {
-        latitude: 52.5896,
-        longitude: 13.28345,
-      },
-    },
-    line: "U6",
-    isHistoric: false,
-  },
-  {
-    timestamp: "2024-03-25T19:24:07.353115Z",
-    station: {
-      id: "U-Hpu",
-      name: "Hermannplatz",
-      coordinates: {
-        latitude: 52.4866057,
-        longitude: 13.424476,
-      },
-    },
-    direction: {
-      id: "SU-HMS",
-      name: "Hermannstraße",
-      coordinates: {
-        latitude: 52.467622,
-        longitude: 13.4309698,
-      },
-    },
-    line: "U8",
-    isHistoric: false,
-  },
-  {
-    timestamp: "2024-03-25T18:56:26.848313Z",
-    station: {
-      id: "SU-Jho",
-      name: "Jungfernheide",
-      coordinates: {
-        latitude: 52.530705,
-        longitude: 13.3004976,
-      },
-    },
-    direction: {
-      id: "U-Rd",
-      name: "Rudow",
-      coordinates: {
-        latitude: 52.4162312,
-        longitude: 13.4952982,
-      },
-    },
-    line: "U7",
-    isHistoric: false,
-  },
-];
